Guard against double-clicking on submissions.

+++ src/pages/PageAddGift.vue
@@ -14,8 +14,8 @@
         <span>Price</span>
         <input v-model="editPrice" type="text" class="form-input__text" />
       </label>
-      <input v-if="gift" class="button button--delete" type="button" @click="deleteGift()" value="Delete" />
-      <input class="button" type="submit" value="Save" />
+      <input v-if="gift" class="button button--delete" type="button" @click="deleteGift()" value="Delete" :disabled="isSaving" />
+      <input class="button" type="submit" value="Save" :disabled="isSaving" />
     </form>
   </div>
 </template>
@@ -38,40 +38,49 @@ export default {
     return {
       editName: this.gift ? this.gift.name : '',
       editLink: this.gift ? this.gift.link : '',
-      editPrice: this.gift ? this.gift.price : ''
+      editPrice: this.gift ? this.gift.price : '',
+      isSaving: false
     }
   },
   methods: {
     async deleteGift () {
-      const userInput = confirm('Are you sure you want to delete?')
-      if (userInput) {
-        const result = await api.deleteGift(this.gift._id)
+      if (this.isSaving === false) {
+        this.isSaving = true
+        const userInput = confirm('Are you sure you want to delete?')
+        if (userInput) {
+          const result = await api.deleteGift(this.gift._id)
+          if (!result.error) {
+            this.$router.push('/mylist')
+          } else {
+            alert(`An error occured: ${result.message}`)
+          }
+        }
+        this.isSaving = false
+      }
+    },
+    async submitForm () {
+      if (this.isSaving === false) {
+        this.isSaving = true
+        let result = null
+        if (this.gift) {
+          result = await api.updateGift(this.gift._id, {
+            name: this.editName,
+            link: this.editLink,
+            price: this.editPrice
+          })
+        } else {
+          result = await api.addGift({
+            name: this.editName,
+            link: this.editLink,
+            price: this.editPrice
+          })
+        }
         if (!result.error) {
           this.$router.push('/mylist')
         } else {
           alert(`An error occured: ${result.message}`)
         }
-      }
-    },
-    async submitForm () {
-      let result = null
-      if (this.gift) {
-        result = await api.updateGift(this.gift._id, {
-          name: this.editName,
-          link: this.editLink,
-          price: this.editPrice
-        })
-      } else {
-        result = await api.addGift({
-          name: this.editName,
-          link: this.editLink,
-          price: this.editPrice
-        })
-      }
-      if (!result.error) {
-        this.$router.push('/mylist')
-      } else {
-        alert(`An error occured: ${result.message}`)
+        this.isSaving = false
       }
     }
   }
